backend/requirements.txt
fastapi
uvicorn
celery
redis
azure-cognitiveservices-speech
yt-dlp
moviepy==1.0.3
numpy
python-multipart